+++ Net12/WebMaze/Views/Shared/_Layout.cshtml
@@ -82,7 +82,7 @@
                         <a class="nav-link" href="/Home/NewCellSugg">New CellSugg</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="/Home/Reviews">Reviews</a>
+                        <a class="nav-link" href="/Reviews/Index">Reviews</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="/CellInfo/AddCell">Add Cell</a>
